requirements.txt
ai-edge-torch
ncnn
pnnx
torchvision==0.20.1
torchaudio==2.5.1
#torch --find-links https://download.pytorch.org/whl/cpu
#torchvision --find-links https://download.pytorch.org/whl/cpu
#torchaudio --find-links https://download.pytorch.org/whl/cpu
